Commit number x

+++ Pinta.Core/Effects/ColorBgra.cs
@@ -377,8 +377,8 @@ namespace Pinta.Core
 
             ret.B =  ClampToByte(Lerp(from.B, to.B, frac));
            
-            ret.G = (byte)ClampToByte(Lerp(from.G, to.G, frac));
-            ret.R = (byte)ClampToByte(Lerp(from.R, to.R, frac));
+            ret.G = ClampToByte(Lerp(from.G, to.G, frac));
+            ret.R = ClampToByte(Lerp(from.R, to.R, frac));
 	    //removing unnesessary cast to byte
